+++ src/app/app.module.ts
@@ -1,6 +1,5 @@
 import { module } from 'angular';
 
 import { DimAppModule } from './dimApp.module';
-import { ShellModule } from './shell/shell.module';
 
-export const AppModule = module('app', [DimAppModule, ShellModule]).name;
+export const AppModule = module('app', [DimAppModule]).name;