+++ src/app/inventory/d2-season-info.ts
@@ -9,7 +9,7 @@ export enum D2SeasonEnum {
 }
 
 // TODO: Update on season change
-export const D2CurrentSeason: number = D2SeasonEnum.BLACK_ARMORY;
+export const D2CurrentSeason: number = D2SeasonEnum.JOKERS_WILD;
 
 export const D2SeasonInfo = {
   1: {
@@ -75,18 +75,18 @@ export const D2SeasonInfo = {
     maxLevel: 50,
     maxPower: 700,
     softCap: 500,
-    releaseDate: '2019-03-05', // TODO: Update this upon confirmation
+    releaseDate: '2019-03-05',
     resetTime: '17:00:00Z'
   },
   7: {
     DLCName: 'Penumbra',
-    seasonName: 'Season of the _______',
+    seasonName: 'Season of Opulence',
     season: 7,
     year: 2,
     maxLevel: 50,
     maxPower: 750,
     softCap: 500,
-    releaseDate: '2019-06-25', // TODO: Update this upon confirmation
+    releaseDate: '2019-06-04', // TODO: Update this upon confirmation
     resetTime: '17:00:00Z'
   }
 };