+++ src/app/search/search-filters.ts
@@ -120,7 +120,7 @@ export function buildSearchConfig(destinyVersion: 1 | 2): SearchConfig {
     'magazine',
     'aimassist',
     'equipspeed',
-    'handling',
+    'handling'
   ];
 
   const source = [
@@ -158,7 +158,8 @@ export function buildSearchConfig(destinyVersion: 1 | 2): SearchConfig {
     'nightfall',
     'adventure',
     'scourge',
-    'blackarmory'
+    'blackarmory',
+    'gambitprime'
   ];
 
   if (destinyVersion === 1) {
@@ -290,7 +291,7 @@ export function buildSearchConfig(destinyVersion: 1 | 2): SearchConfig {
       ikelos: ['ikelos'],
       randomroll: ['randomroll'],
       ammoType: ['special', 'primary', 'heavy'],
-      event: ['dawning', 'crimsondays', 'solstice', 'fotl']
+      event: ['dawning', 'crimsondays', 'solstice', 'fotl', 'revelry']
     });
   }
 
@@ -661,6 +662,10 @@ function searchFilters(
         3650581588,
         3650581589
       ]
+    },
+    gambitprime: {
+      sourceHashes: [1952675042],
+      itemHashes: []
     }
   };